refactor(net): dedup serde str (de)serialize boilerplate via macro
Replace 11 hand-written serde::Serialize/Deserialize impl pairs across the
address, forwarded, socket and uri modules with a single crate-local
impl_serde_str! macro. Two arms cover the two flavors:

- display $t  -> serializer.collect_str(self)   (Display-backed wire form)
- as_str  $t  -> self.as_str().serialize(...)    (borrowed canonical string)

Both deserialize by parsing a Cow<'de, str> through FromStr. Uri previously
called Self::parse(s.as_ref()); its FromStr routes through Uri::parse, so the
macro form is equivalent.

Net -185 lines; behavior unchanged (1087 unit + 43 doctests green).
