+++ rama-net/src/address/domain_address.rs
@@ -156,24 +156,7 @@ impl TryFrom<&[u8]> for DomainAddress {
     }
 }
 
-impl serde::Serialize for DomainAddress {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.collect_str(self)
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for DomainAddress {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let s = <std::borrow::Cow<'de, str>>::deserialize(deserializer)?;
-        s.parse().map_err(serde::de::Error::custom)
-    }
-}
+impl_serde_str!(display DomainAddress);
 
 #[cfg(test)]
 mod tests {
