+++ rama-net/src/address/host_with_opt_port.rs
@@ -601,24 +601,7 @@ impl TryFrom<&[u8]> for HostWithOptPort {
     }
 }
 
-impl serde::Serialize for HostWithOptPort {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.collect_str(self)
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for HostWithOptPort {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let s = <std::borrow::Cow<'de, str>>::deserialize(deserializer)?;
-        s.parse().map_err(serde::de::Error::custom)
-    }
-}
+impl_serde_str!(display HostWithOptPort);
 
 #[cfg(test)]
 mod tests {
