+++ rama-net/src/address/host_with_port.rs
@@ -353,24 +353,7 @@ impl TryFrom<&[u8]> for HostWithPort {
     }
 }
 
-impl serde::Serialize for HostWithPort {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.collect_str(self)
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for HostWithPort {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let s = <std::borrow::Cow<'de, str>>::deserialize(deserializer)?;
-        s.parse().map_err(serde::de::Error::custom)
-    }
-}
+impl_serde_str!(display HostWithPort);
 
 #[cfg(test)]
 mod tests {
