+++ rama-net/src/address/socket_address.rs
@@ -388,24 +388,7 @@ impl TryFrom<&[u8]> for SocketAddress {
     }
 }
 
-impl serde::Serialize for SocketAddress {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.collect_str(self)
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for SocketAddress {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let s = <std::borrow::Cow<'de, str>>::deserialize(deserializer)?;
-        s.parse().map_err(serde::de::Error::custom)
-    }
-}
+impl_serde_str!(display SocketAddress);
 
 #[cfg(test)]
 mod tests {
