+++ rama-net/src/lib.rs
@@ -15,6 +15,9 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![cfg_attr(test, allow(clippy::float_cmp))]
 
+#[macro_use]
+mod macros;
+
 pub mod address;
 pub mod asn;
 pub mod client;
